privategpt/backend/requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - PostgreSQL Support für Railway
sqlalchemy==2.0.25
aiosqlite==0.19.0  # Behalten für lokale Entwicklung
asyncpg==0.29.0  # NEU: Async PostgreSQL Driver
psycopg2-binary==2.9.9  # NEU: PostgreSQL Driver (Fallback)

# Vector Database
chromadb>=0.5.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# LLM & Embeddings - llama-cpp-python für Qwen3-0.6B
httpx==0.27.0  # Behalten für ChromaDB
langchain==0.1.6
langchain-community>=0.0.18
llama-cpp-python==0.2.90  # NEU: Qwen3 Integration (CPU-only)

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Email (Magic-Link)
resend==0.7.0

# Utils
python-dotenv==1.0.0
